Fix: Slow Walk

+++ Osiris/Config.cpp
@@ -99,7 +99,7 @@ void Config::load(size_t id) noexcept
         if (antiAimJson.isMember("LBY Breaker")) antiAim.LBYBreaker = antiAimJson["LBY Breaker"].asBool();
         if (antiAimJson.isMember("LBY Angle")) antiAim.LBYAngle = antiAimJson["LBY Angle"].asFloat();
         if (antiAimJson.isMember("Slow Walk")) antiAim.Slowwalk = antiAimJson["Slow Walk"].asBool();
-        if (antiAimJson.isMember("Slow Walk Speed")) antiAim.slowwalkspeed = antiAimJson["Slow Walk Speed"].asInt();
+        if (antiAimJson.isMember("Slow Walk Speed")) antiAim.slowwalkspeed = antiAimJson["Slow Walk Speed"].asFloat();
         if (antiAimJson.isMember("Slow Walk Key")) antiAim.slowwalkkey = antiAimJson["Slow Walk Key"].asInt();
     }
