+++ Osiris/Config.h
@@ -111,7 +111,7 @@ public:
         bool LBYBreaker{ false };
         float LBYAngle{ 0.0f };
         bool Slowwalk{ false };
-        int slowwalkspeed{ 0 };
+        float slowwalkspeed{ 0.0f };
         int slowwalkkey{ 0 };
     } antiAim;
